% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{is.mlp_net}
\alias{is.mlp_net}
\title{Is it an object of \code{mlp_net} class?}
\usage{
is.mlp_net(x)
}
\arguments{
\item{x}{an object to be checked}
}
\value{
Logical value.
}
\description{
This function checks whether argument is an object of \code{mlp_net} class.
}
\keyword{classes}

