% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{mlp_actvfunc2str}
\alias{mlp_actvfunc2str}
\title{Return character string representing activation function}
\usage{
mlp_actvfunc2str(idx, slope)
}
\arguments{
\item{idx}{activation function index}

\item{slope}{activation function index slope parameter}
}
\value{
This function returns character string representing activation function.
}
\description{
Return character string representing activation function
}
\keyword{internal}

