% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{mlp_check_w}
\alias{mlp_check_w}
\title{Check validity of weight index}
\usage{
mlp_check_w(net, idx = NULL, layer = NULL, nidx = NULL, nplidx = NULL)
}
\arguments{
\item{net}{an object of \code{mlp_net} class}

\item{idx}{integer value, weight absolute index}

\item{layer}{integer value, layer index}

\item{nidx}{integer value, neuron index}

\item{nplidx}{integer value, index of the neuron in the previous layer determining connection
              from neuron \code{nidx} in \code{layer}, 0 denotes bias
              of neuron \code{nidx} in \code{layer}}
}
\value{
This function does not return.
}
\description{
Check validity of weight index
}
\keyword{internal}

