% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{mlp_eval}
\alias{mlp_eval}
\title{Evaluation}
\usage{
mlp_eval(net, input)
}
\arguments{
\item{net}{an object of \code{mlp_net} class}

\item{input}{numeric matrix, each row corresponds to one input vector,
       number of columns must be equal to the number of neurons
       in the network input layer}
}
\value{
Numeric matrix with rows representing network outputs corresponding
        to input rows.
}
\description{
Evaluate network output.
}

