% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_net.R
\docType{methods}
\name{mlp_net-display}
\alias{mlp_net-display}
\alias{print,mlp_net-method}
\alias{show,mlp_net-method}
\alias{summary,mlp_net-method}
\title{Displaying networks (objects of \code{mlp_net} class)}
\usage{
\S4method{show}{mlp_net}(object)

\S4method{print}{mlp_net}(x)

\S4method{summary}{mlp_net}(object)
}
\arguments{
\item{object}{an object of \code{mlp_net} class}

\item{x}{an object of \code{mlp_net} class}
}
\description{
These methods can be used to display objects of \code{mlp_net} class. \code{show}
and \code{print} provide short information about network structure and
activation functions, \code{summary} gives detailed information about
all network connections.
}

