% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{mlp_net-names}
\alias{mlp_get_name}
\alias{mlp_net-names}
\alias{mlp_set_name}
\title{Get and set network names}
\usage{
mlp_get_name(net)

mlp_set_name(net, name)
}
\arguments{
\item{net}{an object of \code{mlp_net} class}

\item{name}{character string with network name}
}
\value{
\code{mlp_get_name} returns character string with network name.

        \code{mlp_set_name} returns network (an object of \code{mlp_net}
              class) with name set to new value.
}
\description{
The following functions can be used for retrieving and setting network names.
}

