% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{mlp_net-weight-indexing}
\alias{mlp_get_w_abs_idx}
\alias{mlp_get_w_idx}
\alias{mlp_net-weight-indexing}
\title{Retrieving absolute weight index}
\usage{
mlp_get_w_idx(net, layer, nidx, nplidx)

mlp_get_w_abs_idx(net, idx)
}
\arguments{
\item{net}{an object of \code{mlp_net} class}

\item{layer}{integer value, layer index}

\item{nidx}{integer value, neuron index}

\item{nplidx}{integer value, index of the neuron in the previous layer determining connection
from neuron \code{nidx} in \code{layer}, 0 denotes bias
of neuron \code{nidx} in \code{layer}}

\item{idx}{integer value, weight index withing active ones}
}
\value{
Absolute weight index.
}
\description{
In some situations weight index (absolute, i.e. within all weights including
inactive ones) needs to be computed based on information about connected
neurons' indices or weight index within actives ones. The latter functionality
is especially useful in implementation of pruning algorithms.
}

