% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{mlp_net-weights-access}
\alias{mlp_get_weights}
\alias{mlp_net-weights-access}
\alias{mlp_set_weights}
\title{Set and retrieve (active) weights' values}
\usage{
mlp_set_weights(net, weights)

mlp_get_weights(net)
}
\arguments{
\item{net}{an object of \code{mlp_net} class}

\item{weights}{numeric vector of new active weights' values}
}
\value{
\code{mlp_set_weights} returns network (an object of \code{mlp_net}
        class) with active weights set to given values.

        \code{mlp_set_weights} returns numeric vector of active weights' values.
}
\description{
One of FCNN's design objectives (and main advantages) is the complete separation
of teaching (and pruning) algorithms from internal network structure workings.
This goal is achieved through fast access to (active) weights vector facilitated
by FCNN's `compressed' network representation. The following two functions
allow users to efficiently retrieve and set network (active) weights vector.
}

