% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlp_plot.R
\name{mlp_plot}
\alias{mlp_plot}
\title{Plotting multilayer perceptron network}
\usage{
mlp_plot(net, show_weights = FALSE)
}
\arguments{
\item{net}{an object of \code{mlp_net} class}

\item{show_weights}{logical, should show weights' values be displayed?
       (FALSE by default)}
}
\value{
This function does not return value.
}
\description{
This function plots a multilayer perceptron network's structure. Optionally,
weights' values are displayed on graph.
}

