% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{mlp_rnd_weights}
\alias{mlp_rnd_weights}
\title{This function sets network weights to random values drawn from uniform
distribution.}
\usage{
mlp_rnd_weights(net, a = 0.2)
}
\arguments{
\item{net}{an object of \code{mlp_net} class}

\item{a}{numeric value, values will be drawn from uniform distribution
on [-a, a] (by default a = 0.2)}
}
\value{
Network (an object of \code{mlp_net} class) with randomised weights.
}
\description{
This function sets network weights to random values drawn from uniform
distribution.
}

