% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{mlp_set_activation}
\alias{mlp_set_activation}
\title{Set network activation functions}
\usage{
mlp_set_activation(net, output = FALSE, activation = c("threshold",
  "sym_threshold", "linear", "sigmoid", "sym_sigmoid", "sigmoid_approx",
  "sym_sigmoid_approx"), slope = NULL)
}
\arguments{
\item{net}{an object of \code{mlp_net} class}

\item{output}{logical value, if TRUE set activation function in the output layer,
if FALSE set activation function in the hidden layer(s)}

\item{activation}{character string, activation function name, admissible
options are: "threshold", "sym_threshold", "linear",
"sigmoid", "sym_sigmoid", "sigmoid_approx", "sym_sigmoid_approx"}

\item{slope}{numeric value, activation function slope parameter}
}
\value{
This function returns network (an object of \code{mlp_net} class)
        with activation function set.
}
\description{
This function sets activation function (and its slope parameter)
for neurons in the hidden layers and in the output layer.
}

