% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{mlp_check_inout}
\alias{mlp_check_inout}
\title{Check validity of inputs and outputs}
\usage{
mlp_check_inout(net, input, output = NULL, i = NULL)
}
\arguments{
\item{net}{an object of \code{mlp_net} class}

\item{input}{numeric matrix, each row corresponds to one input vector,
the number of columns must be equal to the number of neurons
in the network input layer}

\item{output}{numeric matrix with rows corresponding to expected outputs,
the number of columns must be equal to the number of neurons
in the network output layer, the number of rows must be equal to the number
of input rows}

\item{i}{data row index}
}
\value{
This function does not return.
}
\description{
Check validity of inputs and outputs
}
\keyword{internal}

