% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_net.R
\name{mlp_net}
\alias{mlp_net}
\title{Create objects of \code{mlp_net} class}
\usage{
mlp_net(layers, name = NULL)
}
\arguments{
\item{layers}{vector providing numbers of neurons in each layer}

\item{name}{character string, network name (optional)}
}
\value{
Returns an object of \code{mlp_net} class.
}
\description{
Function used for creating multilayer perceptron networks.
}
\examples{

# create a 2-3-1 network
net <- mlp_net(c(2, 3, 1))
# randomise weights
net <- mlp_rnd_weights(net)
# show basic information about the network
show(net)

}
\seealso{
\code{\linkS4class{mlp_net}} for details.
}
\keyword{classes}

