% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_plot.R
\name{mlp_plot}
\alias{mlp_plot}
\title{Plotting multilayer perceptron network}
\usage{
mlp_plot(net, show_weights = FALSE, show_neuron_idx = TRUE)
}
\arguments{
\item{net}{an object of \code{mlp_net} class}

\item{show_weights}{logical, should weights' values be displayed?
(FALSE by default)}

\item{show_neuron_idx}{logical, should neurons' indices be displayed?
(TRUE by default)}
}
\value{
This function does not return value.
}
\description{
This function plots a multilayer perceptron network's structure. Optionally,
weights' values are displayed on graph.
}

